docker/webserver/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
influxdb-client==1.36.1
